requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.10
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.7.0
asgiref==3.8.1
attrs==24.2.0
certifi==2024.8.30
deprecation==2.1.0
Django==5.1.4
django-cors-headers==4.7.0
django-redis==5.4.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
dotenv==0.9.9
drf-spectacular==0.28.0
frozenlist==1.5.0
gotrue==2.11.0
h11==0.13.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.7
httpx==0.27.2
hyperframe==6.0.1
idna==3.10
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
multidict==6.1.0
packaging==24.2
postgrest==0.18.0
propcache==0.2.1
psycopg2-binary==2.9.10
PuLP==2.9.0
pydantic==2.10.3
pydantic_core==2.27.1
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.2
realtime==2.0.6
redis==6.2.0
referencing==0.36.2
rpds-py==0.24.0
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.2
storage3==0.9.0
supabase==2.10.0
supafunc==0.7.0
typing_extensions==4.12.2
uritemplate==4.1.1
websockets==13.1
yarl==1.18.3
